Refactor : console과 web 구분
