+++ src/main/java/org/prgrms/vouchermanagement/webController/HomeController.java
@@ -1,0 +1,12 @@
+package org.prgrms.vouchermanagement.webController;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class HomeController {
+    @GetMapping("/")
+    public String home() {
+        return "home";
+    }
+}
